Disable BLX for now
Currently blx only sets "fake" limits for the charger, which is bypassed
somehow and the battery will continue to charge to 100. I've spent a lot of
time looking at max17048_battery.c to no avail. If anyone has some input,
don't leave me in the dark, but for now I'm setting this aside so I can focus
on caf.